backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
aiofiles
pandas
numpy
scikit-learn
torch
supabase
joblib
